refactor: simplify captured_at stamping
Keep the fix in a small root metadata helper (avoids gitignored lib/*.ts), drop unused timestamp promotion, and slim the unit tests.
